requirements.txt
# BIS GQ Mapper Requirements
# Python packages required for the BIS GQ Mapper application

# Data processing and Excel/CSV handling
pandas>=1.5.0
openpyxl>=3.0.0  # For Excel file support

# YAML configuration file handling
PyYAML>=6.0

# Development and testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0

# Type checking (optional)
mypy>=1.0.0

# Code formatting and linting (optional)
black>=22.0.0
flake8>=5.0.0